edited destroy action in BeersController, displays a message when a beer cannot be destroyed

+++ wood-tavern-backend/app/controllers/beers_controller.rb
@@ -40,7 +40,11 @@ class BeersController < ApplicationController
 
   # DELETE /beers/1
   def destroy
-    beer.destroy
+    beer = Beer.find_by(id: params[:id])
+    if beer
+      beer.destroy
+    else
+      render json: { message: 'Beer not found' }
   end
 
   private
